split helpers to avoid core dependency, fix ts test
