+++ quadratic-client/src/app/ai/tools/endpoint.helper.ts
@@ -1,0 +1,16 @@
+import { AI } from '@/shared/constants/routes';
+import { AIModel } from 'quadratic-shared/typesAndSchemasAI';
+import { isAnthropicModel, isBedrockModel, isOpenAIModel } from './model.helper';
+
+export function getAIProviderEndpoint(model: AIModel, stream: boolean): string {
+  if (isBedrockModel(model)) {
+    return stream ? AI.BEDROCK.STREAM : AI.BEDROCK.CHAT;
+  }
+  if (isAnthropicModel(model)) {
+    return stream ? AI.ANTHROPIC.STREAM : AI.ANTHROPIC.CHAT;
+  }
+  if (isOpenAIModel(model)) {
+    return stream ? AI.OPENAI.STREAM : AI.OPENAI.CHAT;
+  }
+  throw new Error(`Unknown model: ${model}`);
+}
